Move preload and new file menu to separate files
